fixes and simplifications
