+++ .github/workflows/coverage.yml
@@ -29,7 +29,7 @@ jobs:
     - uses: actions/setup-python@v4
       with:
         python-version: 3.11
-    - run: python -m pip install --upgrade pip wheel
+    - run: python -m pip install --upgrade pip
     - run: python -m pip install --prefer-binary coveralls
     - run: python -m pip install --prefer-binary -e .[test]
     - run: coverage run -m pytest
