+++ .github/workflows/docs.yml
@@ -26,7 +26,7 @@ jobs:
       with:
         python-version: "3.9"
     - run: sudo apt-get install pandoc
-    - run: python -m pip install --prefer-binary -v -e .[doc]
+    - run: python -m pip install --prefer-binary -v .[doc]
     - run: python -m ipykernel install --user --name python3
     - run: python docs/build.py
     - uses: actions/upload-pages-artifact@v1
