+++ .github/workflows/test.yml
@@ -39,8 +39,8 @@ jobs:
     - uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
-    - run: python -m pip install --upgrade pip wheel
-    - run: python -m pip install --prefer-binary -v -e .[test]
+    - run: python -m pip install --upgrade pip
+    - run: python -m pip install --prefer-binary -v .[test]
       env:
         DEBUG: 1
     - uses: ts-graphviz/setup-graphviz@v1
