+++ src/pyhepmc/__init__.py
@@ -34,7 +34,6 @@ Missing functionality
   These will be added in the future.
 
 """
-from sys import version_info
 from pyhepmc._core import (
     Units,
     FourVector,
@@ -51,12 +50,6 @@ from pyhepmc._core import (
     equal_particle_sets,
     content,
     listing,
-    _Setup_print_errors,
-    _Setup_set_print_errors,
-    _Setup_print_warnings,
-    _Setup_set_print_warnings,
-    _Setup_debug_level,
-    _Setup_set_debug_level,
     delta_phi,
     delta_eta,
     delta_r2_eta,
@@ -66,15 +59,12 @@ from pyhepmc._core import (
     delta_rap,
 )
 from pyhepmc.io import open as open  # noqa: F401
-import pyhepmc._attributes  # noqa, keep this for its side effects
+from pyhepmc import _attributes
+from pyhepmc._setup import Setup
 from typing import Any
-from importlib.metadata import version, PackageNotFoundError
+from importlib.metadata import version
 
-try:
-    __version__ = version("pyhepmc")
-except PackageNotFoundError:
-    # package is not installed
-    pass
+__version__ = version("pyhepmc")
 
 __all__ = (
     "Units",
@@ -103,46 +93,7 @@ __all__ = (
     "open",
 )
 
-
-class _SetupMeta(type):
-    @property
-    def print_errors(cls) -> bool:
-        "Whether to print errors or not."
-        return _Setup_print_errors()  # type:ignore
-
-    @property
-    def print_warnings(cls) -> bool:
-        "Whether to print warnings or not."
-        return _Setup_print_warnings()  # type:ignore
-
-    @property
-    def debug_level(cls) -> int:
-        "Access debug level."
-        return _Setup_debug_level()  # type:ignore
-
-    def __setattr__(self, name: str, value: Any) -> None:
-        attr = {
-            "print_errors": _Setup_set_print_errors,
-            "print_warnings": _Setup_set_print_warnings,
-            "debug_level": _Setup_set_debug_level,
-        }
-        fn = attr.get(name, None)
-        if fn is None:
-            raise AttributeError
-        fn(value)
-
-
-class Setup(metaclass=_SetupMeta):
-    """
-    Imitates the Setup namespace.
-
-    You can directly read and write to the attributes of this class
-    without creating an instance. They manipulate the corresponding
-    global values in the HepMC3 C++ library.
-    """
-
-    __slots__ = ("print_errors", "print_warnings", "debug_level")
-
+_attributes.install()
 
 try:
     from pyhepmc.view import to_dot
@@ -154,19 +105,6 @@ try:
     GenEvent._repr_html_ = _genevent_repr_html
 except ModuleNotFoundError:  # pragma: no cover
     pass  # pragma: no cover
-
-
-if version_info >= (3, 8):
-    from typing import get_origin, get_args
-else:
-
-    def get_origin(pytype):  # type: ignore  # pragma: no cover
-        if hasattr(pytype, "__origin__"):  # pragma: no cover
-            return pytype.__origin__  # pragma: no cover
-        return None  # pragma: no cover
-
-    def get_args(pytype):  # type: ignore # pragma: no cover
-        return pytype.__args__  # pragma: no cover
 
 
 def __getattr__(name: str) -> Any:
